traider/requirements.txt
yfinance==0.2.31
pandas==2.0.0
numpy==1.24.0
scikit-learn==1.3.0
tensorflow==2.15.0
ta-lib==0.4.28
python-dotenv==1.0.0
pytest==7.4.0